refactor: id를 int에서 long으로 수정

+++ src/main/java/roomescape/controller/ReservationController.java
@@ -39,7 +39,7 @@ public class ReservationController {
 
     @PostMapping
     public ResponseEntity<Reservation> createReservation(@RequestBody ReservationDto reservationDto) {
-        int newIndex = (int) this.index.getAndIncrement();
+        long newIndex = this.index.getAndIncrement();
         Reservation newReservation = new Reservation(newIndex, reservationDto);
         this.reservations.add(newReservation);
