+++ src/main/java/roomescape/model/Reservation.java
@@ -2,8 +2,8 @@ package roomescape.model;
 
 import roomescape.dto.ReservationDto;
 
-public record Reservation(int id, String name, String date, String time) {
-    public Reservation(int id, ReservationDto reservationDto) {
+public record Reservation(long id, String name, String date, String time) {
+    public Reservation(long id, ReservationDto reservationDto) {
         this(id, reservationDto.name(), reservationDto.date(), reservationDto.time());
     }
